+++ tests/fixtures/services/illumina_services/metrics/230622_A00621_0864_AHY7FFDRX2/Unaligned/Reports/Demultiplex_Stats.csv
