+++ tests/services/illumina_services/demux_version/test_illumina_demux_version_service.py
@@ -2,7 +2,7 @@
 
 from pathlib import Path
 
-from cg.services.illumina_services.illumina_metrics_service.illumina_demux_version_service import (
+from cg.services.illumina_services.metrics.demux_version_service import (
     IlluminaDemuxVersionService,
 )
