+++ tests/services/illumina_services/metrics/parsers/test_bclconvert_metrics_parser.py
@@ -4,11 +4,11 @@ from pathlib import Path
 
 import pytest
 
-from cg.services.illumina_services.illumina_metrics_service.models import (
+from cg.services.illumina_services.metrics.models import (
     SequencingQualityMetrics,
     DemuxMetrics,
 )
-from cg.services.illumina_services.illumina_metrics_service.bcl_convert_metrics_parser import (
+from cg.services.illumina_services.metrics.bcl_convert_metrics_parser import (
     BCLConvertMetricsParser,
 )
