+++ tests/store/crud/add/test_store_add_base.py
@@ -3,7 +3,7 @@ from datetime import datetime as dt
 import pytest
 
 from cg.constants.subject import Sex
-from cg.services.illumina_services.illumina_metrics_service.models import IlluminaFlowCellDTO
+from cg.services.illumina_services.metrics.models import IlluminaFlowCellDTO
 from cg.store.models import (
     ApplicationVersion,
     Collaboration,
