+++ tests/store/filters/test_status_illumina_flow_cell_filters.py
@@ -3,7 +3,7 @@
 from sqlalchemy.orm import Query
 
 from cg.models.run_devices.illumina_run_directory_data import IlluminaRunDirectoryData
-from cg.services.illumina_services.illumina_metrics_service.models import IlluminaFlowCellDTO
+from cg.services.illumina_services.metrics.models import IlluminaFlowCellDTO
 from cg.store.filters.status_illumina_flow_cell_filters import (
     filter_illumina_flow_cell_by_internal_id,
 )
